Clean up A* logic
This change cleans up some A* logic. The neighbors function might need to find the closest neighbors via a Bucket fill kind of algorithm as opposed to adjacent neighbors. Since there is a space around each of the obstacles that the enemy can't traverse into, it's possible that it will get stuck and be unable to move (or the algorithm will get stuck in an infinite loop since no neighbors are found).

* Update GUI.java
* Update Player.java
* Update Node.java
* Update Pathfinding.java