Fix most of the issues with A*
I added an Obstacle object which enables the traversability of nodes to be more precisely calculated. While this does a better job of preventing issues with "incomputable" paths, I also check that the start and target are both traversable nodes to prevent the possibility that the path will never be found and will cause expensive computation.

* Update GUI.java
* Add Obstacle.java
* Update Pathfinding.java